un-inline finishing touches power
