+++ src/main/resources/fishingResources/localization/kor/Cardstrings.json
@@ -330,13 +330,7 @@
   },
   "fishing:FinishingTouches": {
     "NAME": "마무리 일격",
-    "DESCRIPTION": "카드를 2 장 버립니다. NL 카드를 사용했을때, 그 카드가 손패에 유일하게 남아있는 카드라면, 2 번 사용합니다.",
-    "EXTENDED_DESCRIPTION": [
-      "마무리 일격",
-      "카드를 사용했을때, 그 카드가 손패에 유일하게 남아있는 카드라면, 2 번 사용합니다.",
-      "카드를 사용했을때, 그 카드가 손패에 유일하게 남아있는 카드라면, 그 카드를 추가로 #b",
-      " 번 사용합니다."
-    ]
+    "DESCRIPTION": "카드를 2 장 버립니다. NL 카드를 사용했을때, 그 카드가 손패에 유일하게 남아있는 카드라면, 2 번 사용합니다."
   },
   "fishing:FlyingFish": {
     "NAME": "날치",
