+++ src/main/resources/fishingResources/localization/kor/Powerstrings.json
@@ -70,5 +70,13 @@
       "턴 종료시, #y포일 카드를 2 장 이상 사용했었다면, #y방어도를 !M! #b",
       " 얻습니다."
     ]
+  },
+  "fishing:FinishingTouchesPower": {
+    "NAME": "마무리 일격",
+    "DESCRIPTIONS": [
+      "카드를 사용했을때, 그 카드가 손패에 유일하게 남아있는 카드라면, 2 번 사용합니다.",
+      "카드를 사용했을때, 그 카드가 손패에 유일하게 남아있는 카드라면, 그 카드를 추가로 #b",
+      " 번 사용합니다."
+    ]
   }
 }
